test: add tests for missing code branches
