+++ frontend/hooks/listen-for-active-tab/listen-for-active-tab.spec.ts
@@ -1,12 +1,12 @@
 import { renderHook } from '@testing-library/react'
 
-import { useListenForPopups } from '.'
+import { useListenForActiveTab } from '.'
 
 const setup = jest.fn()
 const teardown = jest.fn()
 
-jest.mock('@/stores/popover-store', () => ({
-  usePopoverStore: jest.fn().mockImplementation((function_) =>
+jest.mock('@/stores/tab-store', () => ({
+  useTabStore: jest.fn().mockImplementation((function_) =>
     function_({
       setup,
       teardown
@@ -14,9 +14,9 @@ jest.mock('@/stores/popover-store', () => ({
   )
 }))
 
-describe('ListenForPopups', () => {
+describe('ListenForActiveTab', () => {
   it('sets up and tears down listeners appropriately', async () => {
-    const view = renderHook(() => useListenForPopups())
+    const view = renderHook(() => useListenForActiveTab())
     expect(setup).toHaveBeenCalled()
     expect(teardown).not.toHaveBeenCalled()
     view.unmount()
