+++ frontend/hooks/listen-for-popups/listen-for-popups.ts
@@ -1,0 +1,14 @@
+import { useEffect } from 'react'
+
+import { usePopoverStore } from '@/stores/popover-store'
+
+export function useListenForPopups() {
+  const { setup, teardown } = usePopoverStore((state) => ({
+    setup: state.setup,
+    teardown: state.teardown
+  }))
+  useEffect(() => {
+    setup()
+    return teardown
+  }, [setup, teardown])
+}
